+++ app/scenes/myJobs/Past.js
@@ -3,7 +3,7 @@ import { Container, Row, Col, Table } from "reactstrap";
 import PropTypes from "prop-types";
 import moment from "moment";
 import Nav from "../dashboard/components/Nav";
-import { firestore } from "../../../app/firebase";
+import { firestore, auth } from "../../firebase";
 import { collectIdsAndDocs, formatDate } from "../../../app/utilities";
 
 const DashTable = props => {
@@ -40,10 +40,10 @@ const DashTable = props => {
 const FullTable = props => {
   const { jobs, handleRemove, handleEdit } = props;
   return (
-    <Container fluid>
-      <Row>
+    <div className="container-fluid">
+      <div className="row">
         <Nav />
-        <Col>
+        <div className="col content shadow">
           <h5 className="tableHeading">Past Jobs</h5>
           <Table hover striped responsive>
             <thead>
@@ -85,9 +85,9 @@ const FullTable = props => {
               })}
             </tbody>
           </Table>
-        </Col>
-      </Row>
-    </Container>
+        </div>
+      </div>
+    </div>
   );
 };
 
@@ -102,15 +102,19 @@ export default class Past extends React.Component {
   unsubscribe = null;
 
   componentDidMount = async () => {
+    this.getJobsData();
+  };
+
+  getJobsData = async () => {
     this.unsubscribe = await firestore
-      .collection("jobs")
+      .collection(`users/${this.props.user.uid}/jobs`)
       .onSnapshot(snapshot => {
         const jobs = snapshot.docs.map(collectIdsAndDocs);
-        this.setState({ jobs }, console.log(this.state));
+        this.setState({ jobs });
       });
   };
 
-  componentWillUnmount = () => {
+  componentWillUnmount = async () => {
     this.unsubscribe();
   };
 
@@ -121,13 +125,15 @@ export default class Past extends React.Component {
   };
 
   handleRemove = async id => {
-    firestore.doc(`jobs/${id}`).delete();
+    firestore.doc(`users/${this.props.user.uid}jobs/${id}`).delete();
   };
-  static propTypes = { fullView: PropTypes.bool.isRequired };
-  static defaultProps = { fullView: true };
+
   render() {
+    console.log(auth.currentUser);
     if (!this.state.jobs) {
-      return <div>loading</div>;
+      return (
+        <small className="text-muted">You have no past jobs to display</small>
+      );
     } else {
       return (
         <Col>
